backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.5.0
python-multipart==0.0.20
email-validator==2.3.0

# MySQL async drivers
sqlalchemy==2.0.23
asyncmy==0.2.9
aiomysql==0.2.0
greenlet==3.0.3

# Utilities
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click>=8.0.0,<8.2.0
cryptography>=40.0.0
h11==0.16.0
idna==3.10
requests>=2.31.0
sniffio==1.3.1
starlette==0.37.2
urllib3>=2.0.0
watchfiles>=0.20.0
